Move error elements out of instruction module
There were several error components in the instruction module that arguably didn't belong there.

This moves `InstructionError` into its own file, and the `MapInstructionError` trait into the the `error` moduel (where `InstructionResult` is defined).

I also realized that there aren't any boolean instruction errors, and I don't know what boolean errors there might be, so I deleted `BoolInstructionError`.
