+++ packages/push/src/error/mod.rs
@@ -16,6 +16,33 @@ pub enum Error<S, E> {
 
 pub type InstructionResult<S, E> = core::result::Result<S, Error<S, E>>;
 
+/// Maps a (presumably error) type into an `InstructionResult`.
+/// This is used to convert `InstructionResult<S, E1>`
+/// into `InstructionResult<S, E2>`, i.e. do `map_err()` on
+/// the inner error types of an `InstructionResult`, preserving
+/// the other fields in `Error`.
+pub trait MapInstructionError<S, E> {
+    ///
+    /// # Errors
+    ///
+    /// This always returns an error type.
+    fn map_err_into(self) -> InstructionResult<S, E>;
+}
+
+// MizardX@Twitch's initial suggestion here had `E2` as a generic on the
+// _function_ `map_err_into()` instead of at the `impl` level. That provided
+// some additional flexibility, although it wasn't clear that we would use it.
+// The current approach (suggested by esitsu@Twitch) simplified the
+// `MapInstructionError` trait in a nice way, so I went with that.
+impl<S, E1, E2> MapInstructionError<S, E2> for InstructionResult<S, E1>
+where
+    E1: Into<E2>,
+{
+    fn map_err_into(self) -> InstructionResult<S, E2> {
+        self.map_err(|e| e.map_inner_err(Into::into))
+    }
+}
+
 impl<S, E> Error<S, E> {
     pub fn fatal(state: S, error: impl Into<E>) -> Self {
         Self::Fatal(FatalError::new(state, error.into()))
